fix modal use
